+++ primary/src/primary.rs
@@ -65,9 +65,9 @@ pub enum ExecutorPrimaryMessage {
 #[derive(Debug, Serialize, Deserialize)]
 pub enum WorkerPrimaryMessage {
     /// The worker indicates it sealed a new batch.
-    OurBatch(Digest, WorkerId, BTreeMap<u64, u64>, QuorumMetrics),
+    OurBatch(Digest, WorkerId, BTreeMap<u32, u16>, QuorumMetrics),
     /// The worker indicates it received a batch's digest from another authority.
-    OthersBatch(Digest, WorkerId, BTreeMap<u64, u64>),
+    OthersBatch(Digest, WorkerId, BTreeMap<u32, u16>),
     /// Replay mode: worker notifies that a batch is generated and stored (batch_index, digest, worker_id).
     ReplayBatchReady(u64, Digest, WorkerId),
 }
@@ -125,7 +125,7 @@ impl Primary {
         let secret = keypair.secret;
 
         let (tx_others_digests, rx_others_digests) = channel(CHANNEL_CAPACITY);
-        let (tx_our_digests, rx_our_digests): (Sender<(Digest, WorkerId, BTreeMap<u64, u64>, QuorumMetrics)>, _) = channel(CHANNEL_CAPACITY);
+        let (tx_our_digests, rx_our_digests): (Sender<(Digest, WorkerId, BTreeMap<u32, u16>, QuorumMetrics)>, _) = channel(CHANNEL_CAPACITY);
         let (tx_parents, rx_parents) = channel(CHANNEL_CAPACITY);
         let (tx_headers, rx_headers) = channel(CHANNEL_CAPACITY);
         let (tx_sync_headers, rx_sync_headers) = channel(CHANNEL_CAPACITY);
@@ -299,8 +299,8 @@ impl MessageHandler for PrimaryReceiverHandler {
 /// Defines how the network receiver handles incoming workers messages.
 #[derive(Clone)]
 struct WorkerReceiverHandler {
-    tx_our_digests: Option<Sender<(Digest, WorkerId, BTreeMap<u64, u64>, QuorumMetrics)>>,
-    tx_others_digests: Option<Sender<(Digest, WorkerId, BTreeMap<u64, u64>)>>,
+    tx_our_digests: Option<Sender<(Digest, WorkerId, BTreeMap<u32, u16>, QuorumMetrics)>>,
+    tx_others_digests: Option<Sender<(Digest, WorkerId, BTreeMap<u32, u16>)>>,
     tx_replay_ready: Option<Sender<(u64, Digest, WorkerId)>>,
 }
